Edit option in ManageCategories is fully Functional

+++ myAmazon-v1/AdminPanel/AddNewBrand.aspx.cs
@@ -38,23 +38,72 @@ namespace myAmazon_v1.AdminPanel
                 id_category_name.DataValueField = "id";
                 id_category_name.DataBind();
                 id_category_name.Items.Insert(0, new ListItem("--Select--", "NA"));
+
+                if(Session["isEdit"] != null && Session["isEdit"].Equals("1"))
+                {
+                    this.Title = "Edit Brand";
+                    id_brand_title.Text = "Edit Brand";
+                    id_submit_brand.Text = "Update";
+
+                    cmd = "SELECT Name AS name, [Desc] AS [desc], CategoryId AS catId FROM Brand WHERE id=" + Session["BrandId"];
+                    sqlCmd = new SqlCommand(cmd, conn);
+                    SqlDataReader reader = null;
+
+                    try
+                    {
+                        conn.Open();
+                        reader = sqlCmd.ExecuteReader();
+                        reader.Read();
+                        id_brand_name.Text = reader["name"].ToString();
+                        id_category_name.SelectedValue = reader["catId"].ToString();
+                        StreamReader stream = new StreamReader(Server.MapPath(reader["desc"].ToString()));
+                        id_brand_desc.Text = stream.ReadToEnd();
+                        stream.Close();
+                        conn.Close();
+                    }
+                    catch (Exception ex)
+                    {
+                        id_log_brand.Text += ex.ToString();
+                        conn.Close();
+                    }
+                }
             }
         }
 
         protected void Press_Submit(object sender, EventArgs e)
         {
             SqlConnection conn = new SqlConnection(@"Data Source=DESKTOP-HO7NA1P;Initial Catalog=myAmazon;User ID=sa;Password=root");
-            SqlCommand insert = new SqlCommand("INSERT INTO Brand(Name, [CategoryId]) OUTPUT inserted.id VALUES(@name, @category)", conn);
-            insert.Parameters.AddWithValue("@name", id_brand_name.Text);
-            insert.Parameters.AddWithValue("@category", id_category_name.SelectedValue);
+            string cmd = "";
+            bool isEdit = false;
+
+            if (Session["isEdit"] != null && Session["isEdit"].Equals("1"))
+                isEdit = true;
+
+            if(isEdit)
+                cmd = "UPDATE Brand SET Name=@name, [CategoryId]=@category WHERE id=" + Session["BrandId"];
+            else
+                cmd = "INSERT INTO Brand(Name, [CategoryId]) OUTPUT inserted.id VALUES(@name, @category)";
+
+            SqlCommand sqlCmd = new SqlCommand(cmd, conn);
+            sqlCmd.Parameters.AddWithValue("@name", id_brand_name.Text);
+            sqlCmd.Parameters.AddWithValue("@category", id_category_name.SelectedValue);
             int id = 0;
             try
             {
                 conn.Open();
-                id = (int )insert.ExecuteScalar();
+                if(isEdit)
+                {
+                    id = Convert.ToInt32(Session["BrandId"]);
+                    sqlCmd.ExecuteNonQuery();
+                    id_log_brand.Text = "Successfully Updated";
+                } 
+                else
+                {
+                    id = (int)sqlCmd.ExecuteScalar();
+                    id_brand_name.Text = "";
+                    id_log_brand.Text = "Successfully Inserted";
+                }
                 conn.Close();
-                id_log_brand.Text = "Successfully Inserted";
-                id_brand_name.Text = "";
             }
             catch (Exception ex)
             {
@@ -73,13 +122,23 @@ namespace myAmazon_v1.AdminPanel
                 {
                     id_log_brand.Text += ex.ToString();
                 }
-                File.Create(Server.MapPath("~/BrandsData/" + id.ToString() + ".txt")).Close();
+                if(!File.Exists(Server.MapPath("~/BrandsData/" + id.ToString() + ".txt")))
+                    File.Create(Server.MapPath("~/BrandsData/" + id.ToString() + ".txt")).Close();
                 File.WriteAllText(Server.MapPath("~/BrandsData/" + id.ToString() + ".txt"), id_brand_desc.Text);
-                id_brand_desc.Text = "";
-                conn.Open();
-                SqlCommand query = new SqlCommand("UPDATE Brand SET [Desc] ='" + "~/BrandsData/" + id.ToString() + ".txt" + "' WHERE id=@cid", conn);
-                query.Parameters.AddWithValue("@cid", id);
-                query.ExecuteNonQuery();
+                if (isEdit)
+                {
+                    Session["isEdit"] = "0";
+                    Response.Redirect(@"..\AdminPanel\ManageBrands.aspx");
+                }
+                else
+                {
+                    id_brand_desc.Text = "";
+                    conn.Open();
+                    SqlCommand query = new SqlCommand("UPDATE Brand SET [Desc] ='" + "~/BrandsData/" + id.ToString() + ".txt" + "' WHERE id=@cid", conn);
+                    query.Parameters.AddWithValue("@cid", id);
+                    query.ExecuteNonQuery();
+                    conn.Close();
+                }
             }
             catch (Exception ex)
             {
